Animate decrease in progress
Fixed an error causing regression (declining progression) to
animate from 0 instead of the last value.
This happened regardless if the `animateFromLastPercent` was set
too true or not.

The error was an unnecessary logical operation checking if the new value
was larger than the old.
Removing this statement fixed the bug and does not effect any other code.