Add border radius to the linear progress bar
In the case, the `llinearStrokeCap` parameter will be deprecated.

Uses path.addRRect() instead of drawLine() to draw the progress bar

Allows user to control the border radius of the progress bar.
The progress bar will always have the same ends at the moment, will improve that in the future.

The basic functionality are all there, will refactor some code after a while.